Fix: 확인 버튼 활성화 관련 파일 나누기 및 분류하기
submitBtnHandler와 관련된 함수들이 들어있던 파일을 작은 함수들로 나누고
분류를 다시 했다.
