+++ src/js/components/inputbar/activateBtn.js
@@ -1,0 +1,6 @@
+export function activateBtn() {
+	const submitBtn = document.querySelector('.submit-btn svg');
+	submitBtn.setAttribute('style', 'cursor: pointer');
+	submitBtn.children[0].setAttribute('fill', '#F5B853');
+	submitBtn.children[0].setAttribute('fill-opacity', '1');
+}
