+++ src/js/index.js
@@ -1,15 +1,15 @@
-import { initInputDate } from './components/inputbar/initInputDate.js';
+import { initInputBarDate } from './components/inputbar/initInputBarDate.js';
 import { initHeaderMonthYear } from './components/header/initHeaderMonthYear.js';
+import { initFormEventChecker } from './components/inputbar/initFormChecker.js';
 import './components/inputbar/dropdown.js';
-import './utils/formValidation.js';
 import './components/inputbar/autoComma.js';
 import './components/inputbar/switchCategory.js';
 import './data/formData.js';
-import './utils/submitBtnHandelr.js';
 
 const init = () => {
 	initHeaderMonthYear();
-	initInputDate();
+	initInputBarDate();
+	initFormEventChecker();
 };
 
 window.addEventListener('load', init);
